chore(Android): Refactor out nearby page logic (#981)
